+++ src/frame/data.rs
@@ -148,7 +148,7 @@ impl<T: Buf> Data<T> {
     ///
     /// Panics if `dst` cannot contain the data frame.
     pub(crate) fn encode_chunk<U: BufMut>(&mut self, dst: &mut U) {
-        let len = self.data.remaining() as usize;
+        let len = self.data.remaining();
 
         assert!(dst.remaining_mut() >= len);