+++ src/proto/settings.rs
@@ -1,6 +1,5 @@
 use crate::codec::UserError;
 use crate::error::Reason;
-use crate::frame;
 use crate::proto::*;
 use std::task::{Context, Poll};
 
@@ -58,6 +57,10 @@ impl Settings {
 
                     if let Some(max) = local.max_header_list_size() {
                         codec.set_max_recv_header_list_size(max as usize);
+                    }
+
+                    if let Some(val) = local.header_table_size() {
+                        codec.set_recv_header_table_size(val as usize);
                     }
 
                     streams.apply_local_settings(local)?;